Add get_cached_list and get_obj_by_field
Add get_cached_list, which gets the list of all users/forums/teams from
a JSON file if it exists, or gets it from Ryver if it does not.

Add get_obj_by_field and a bunch of fields, which simply gets the
user/forum/team with a field matching a value.