Add function to react to message as well as supporting functions
By using the Chat.React() API function, we are able to react to
messages. This requires both the chat ID as well as the message ID's,
which spawned the get_chat_id() and get_chat() functions which return
the ID of the chat or a Chat object, respectively. get_chat_type() is
also needed to ensure compatibility with different types of chats. A
top-level function outside of any classes, get_type_from_entity() is
also used to help react().