Improve notifications and add new message function
Add Chat.get_message_from_id(), which gets a message (and some before
and after it) from an ID.

Delete the old methods in Notification and add new methods and clean up
the documentation. Although the purpose of the subject, object and via
of a notification is unclear, the new methods should be good for most
purposes.

Add a new method to Notification that allows setting a notification as
read/unread or seen/unseen.