Improve API, refactor and add methods to Object
Improve parts of the Topics API to make it more consistent in terms of
which methods are coroutines and which ones are normal functions.

Add get_create_user() and get_modify_user() methods to Object.

Add Object.get_deferred_field() and refactor the code to make some
methods use the new method.