Fix infinite loop on errors, deprecate RyverWS.on_error
Fix the realtime API getting stuck in a loop of "Unexpected binary
message received: X is not str" when the connection dies unexpectedly
and isn't detected by the pings. Deprecate on_error, as the connection
will be closed automatically or recovered on errors.

Errors and warnings from the websocket will now be written to stderr.
Reduce log output by approximately 99.999852% so we will no longer get
257874670-byte log files.